+++ src/actions/create_transfer.ts
@@ -21,23 +21,24 @@ interface QoyodTransfer {
 }
 
 export const create_transfer = async (event: EVENT, options: Config) => {
-  console.log("*".repeat(20));
-
   const repzo = new Repzo(options.data?.repzoApiKey, { env: options.env });
   const action_sync_id: string = event?.headers?.action_sync_id || uuid();
   const actionLog = new Repzo.ActionLogs(repzo, action_sync_id);
   let body: Service.Transfer.Schema | any;
   try {
-    console.log("create_transfer");
+    // console.log("create_transfer");
     await actionLog.load(action_sync_id);
-    await actionLog.addDetail(`Repzo Qoyod: Started Create Transfer`).commit();
 
     body = event.body;
     try {
       if (body) body = JSON.parse(body);
     } catch (e) {}
 
-    const result = { created: 0, failed: 0, failed_msg: [] };
+    await actionLog
+      .addDetail(
+        `Repzo Qoyod: Started Create Transfer - ${body?.serial_number?.formatted}`
+      )
+      .commit();
 
     const repzo_transfer = body;
 
@@ -103,18 +104,21 @@ export const create_transfer = async (event: EVENT, options: Config) => {
       },
     };
 
-    console.dir(qoyod_transfer_body, { depth: null });
+    // console.dir(qoyod_transfer_body, { depth: null });
 
-    const qoyod_transfer = await _create(
+    const result = await _create(
       options.serviceEndPoint,
       "/inventory_transfers",
       qoyod_transfer_body,
       { "API-KEY": options.data.serviceApiKey }
     );
 
-    console.log(qoyod_transfer);
-    console.log(result);
-    await actionLog.setStatus("success", result).setBody(body).commit();
+    // console.log(result);
+    await actionLog
+      .setStatus("success", result)
+      .setBody(body)
+      .setMeta(qoyod_transfer_body)
+      .commit();
     return result;
   } catch (e: any) {
     //@ts-ignore