+++ src/actions/index.ts
@@ -2,18 +2,21 @@ import { Config, Action } from "../types";
 import { EVENT } from "../types";
 import { create_invoice } from "./create_invoice.js";
 import { create_payment } from "./create_payment.js";
-import { create_transfer } from "./create_transfer.js";
+// import { create_transfer } from "./create_transfer.js";
 import { create_client } from "./create_client.js";
 import { create_refund } from "./create_refund.js";
+import { create_creditNote } from "./create_return_invoice.js";
 
 export const actions = async (event: any, options: Config) => {
   switch (event.queryStringParameters?.action) {
     case "create_invoice":
       return await create_invoice(event, options);
+    case "create_creditNote":
+      return await create_creditNote(event, options);
     case "create_payment":
       return await create_payment(event, options);
-    case "create_transfer":
-      return await create_transfer(event, options);
+    // case "create_transfer":
+    // return await create_transfer(event, options);
     case "create_client":
       return await create_client(event, options);
     case "create_refund":
@@ -30,15 +33,20 @@ export const actionsList: Action[] = [
     description: "create invoice ..",
   },
   {
+    action: "create_creditNote",
+    name: "create credit note",
+    description: "create credit note ..",
+  },
+  {
     action: "create_payment",
     name: "create payment",
     description: "create payment ..",
   },
-  {
-    action: "create_transfer",
-    name: "create transfer",
-    description: "create transfer ..",
-  },
+  // {
+  //   action: "create_transfer",
+  //   name: "create transfer",
+  //   description: "create transfer ..",
+  // },
   {
     action: "create_client",
     name: "create client",